API v1 : Implemented cached data for trending songs, added scheduler and network manager utility.
The data for trending songs is now precached and stored in SQLite database. For handling it, helpers.database is added.

helpers.networking is added for better handling of page load requests that are made from the server.

Scheduler class is added, which can be used to add scheduled processes easily.